npm: regenerate package-lock.json for npm v7+
I had intention to install locked dependencies and called `npm install`
in the project directory first time. Surprisingly, it updates my
`package-lock.json` with `lockfileVersion: 2` (was 1) and the new
`packages` field.

The documentation [1] says that it is the new lock file format for npm
v7+, which is compatible with npm v6. Okay, let's move to the new
format.

FWIW, my versions are the following:

| Package | Version |
| ------- | ------- |
| nodejs  | 16.13.1 |
| npm     | 8.1.2   |

[1]: https://docs.npmjs.com/cli/v8/configuring-npm/package-lock-json