build: validate that an error is of type Error
`npm run pre-checkin` (more precisely `tsc`) says the following:

 | src/main.ts:178:20 - error TS2571: Object is of type 'unknown'.
 |
 | 178       core.warning(error.message)
 |                        ~~~~~
 |
 | src/main.ts:184:20 - error TS2571: Object is of type 'unknown'.
 |
 | 184     core.setFailed(error.message)
 |                        ~~~~~
 |
 |
 | Found 2 errors.

In the previous commit I called `npm update` and it updates TypeScript
from 4.1.2 to 4.5.5. Since 4.4 it requires to explicitly check that a
raised error is of type Error, when the strict mode is enabled (see
[1]). We have the strict mode enabled in `tsconfig.json`.

I also found that TypeScript 4.0 allows to mark a raised object as
`unknown` explicitly (see [2]) and I considered it as the recommendation
to apply.

The similar change was applied in the TypeScript Action template (see
[3]).

[1]: https://www.typescriptlang.org/docs/handbook/release-notes/typescript-4-4.html#using-unknown-in-catch-variables
[2]: https://www.typescriptlang.org/docs/handbook/release-notes/typescript-4-0.html#unknown-on-catch-clause-bindings
[3]: https://github.com/actions/typescript-action/issues/376